Create student quizzes from full quiz sheets

This function takes all complete quiz Excel sheets. It checks each quiz sheet for proper formatting. It then strips some of the content and produces quiz sheets to be given to students for filling out. It places those quiz sheets into the student_quiz_sheets folder

create_studentquizzes(courselocation)

Arguments

courselocation

Path to main course directory

Value

errormsg Null if quiz generation was successful, otherwise contains an error message